+++ ESP32-S3-DevKitC1-Xtensa-FreeRTOS-bsp/projects/microej/platform/inc/LLSEC_ERRORS.h
@@ -13,8 +13,8 @@
  * @file
  * @brief MicroEJ Security low level API
  * @author MicroEJ Developer Team
- * @version 2.3.0
- * @date 9 November 2023
+ * @version 2.4.0
+ * @date 16 February 2024
  */
 
 #define J_SEC_NO_ERROR							 0 /* not an error */