+++ ESP32-S3-DevKitC1-Xtensa-FreeRTOS-bsp/projects/microej/platform/inc/LLUI_INPUT_impl.h
@@ -1,5 +1,5 @@
 /* 
- * Copyright 2012-2022 MicroEJ Corp. All rights reserved.
+ * Copyright 2012-2023 MicroEJ Corp. All rights reserved.
  * This library is provided in source code for use, modification and test, subject to license terms.
  * Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
  */
@@ -15,7 +15,11 @@ extern "C" {
  * of API to connect the MicroUI input pump. All APIs are mandatory.
  */
 
-#include "LLUI_INPUT.h"
+// --------------------------------------------------------------------------------
+// Includes
+// --------------------------------------------------------------------------------
+
+#include <LLUI_INPUT.h>
 
 // --------------------------------------------------------------------------------
 // Constants
@@ -24,12 +28,12 @@ extern "C" {
 /*
  * @brief Low-Level API UI major version.
  */
-#define LLUI_MAJOR_VERSION 13
+#define LLUI_MAJOR_VERSION 14
 
 /*
  * @brief Low-Level API UI minor version.
  */
-#define LLUI_MINOR_VERSION 7
+#define LLUI_MINOR_VERSION 0
 
 /*
  * @brief Low-Level API UI patch version.