+++ ESP32-S3-DevKitC1-Xtensa-FreeRTOS-bsp/projects/microej/platform/inc/intern/LLSEC_CIPHER_impl.h
@@ -1,7 +1,7 @@
 /*
  * C
  *
- * Copyright 2017-2021 MicroEJ Corp. All rights reserved.
+ * Copyright 2017-2024 MicroEJ Corp. All rights reserved.
  * This library is provided in source code for use, modification and test, subject to license terms.
  * Any modification of the source code will break MicroEJ Corp. warranties on the whole library.
  */
@@ -10,13 +10,14 @@
  * @file
  * @brief MicroEJ Security low level API
  * @author MicroEJ Developer Team
- * @version 2.3.0
- * @date 9 November 2023
+ * @version 2.4.0
+ * @date 16 February 2024
  */
 
 #define LLSEC_CIPHER_IMPL_get_transformation_description		Java_com_is2t_support_security_cipher_NativeCipherSpi_nativeGetTransformationDescription
 #define LLSEC_CIPHER_IMPL_get_buffered_length					Java_com_is2t_support_security_cipher_NativeCipherSpi_nativeGetBufferedLength
 #define LLSEC_CIPHER_IMPL_get_IV								Java_com_is2t_support_security_cipher_NativeCipherSpi_nativeGetIV
+#define LLSEC_CIPHER_IMPL_set_IV								Java_com_is2t_support_security_cipher_NativeCipherSpi_nativeSetIV
 #define LLSEC_CIPHER_IMPL_get_IV_length							Java_com_is2t_support_security_cipher_NativeCipherSpi_nativeGetIVLength
 #define LLSEC_CIPHER_IMPL_init									Java_com_is2t_support_security_cipher_NativeCipherSpi_nativeInit
 #define LLSEC_CIPHER_IMPL_decrypt								Java_com_is2t_support_security_cipher_NativeCipherSpi_nativeDecrypt