+++ ESP32-S3-DevKitC1-Xtensa-FreeRTOS-bsp/projects/microej/platform/inc/intern/LLSEC_KEY_PAIR_GENERATOR_impl.h
@@ -10,8 +10,8 @@
  * @file
  * @brief MicroEJ Security low level API
  * @author MicroEJ Developer Team
- * @version 2.3.0
- * @date 9 November 2023
+ * @version 2.4.0
+ * @date 16 February 2024
  */
 
 #define LLSEC_KEY_PAIR_GENERATOR_IMPL_get_algorithm 					Java_com_microej_support_security_keypairgen_NativeKeyPairGeneratorSpi_nativeGetAlgorithm